OAK-10973  separate DocumentPropertyState from CompressedDocumentPropertyState refactor tests
